answers view created, logic for view moved to it. Changes to be committed: 	modified:   app/views/__init__.py 	copied:     app/views/__init__.py -> app/views/answers.py
